requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-elasticsearch-dsl==7.3
django-environ==0.10.0
django-filter==23.2
django-material-admin==1.8.6
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
fcm-django==2.0.0
firebase-admin==6.2.0
frozenlist==1.4.0
geoip2==4.7.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
inflection==0.5.1
maxminddb==2.4.0
msgpack==1.0.5
multidict==6.0.4
packaging==23.1
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.24.0
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
